Convert models/colors
