+++ src/@types/random-hex-color.d.ts
@@ -1,0 +1,3 @@
+declare module 'random-hex-color' {
+  export default function randomColor(): `#${string}`;
+}
